Move to mixin class
